requirements.txt
mtcnn
numpy==1.19.5
Pillow==8.2.0
keras==2.4.3
tensorflow==2.5.0
tensorflow-cpu==2.5.0
pandas==1.2.4
sklearn==0.0
matplotlib==3.4.2
setuptools==57.0.0
opencv-python==1.19.5
keras-nightly==2.5.0.dev2021032900
tensorflow-estimator==2.5.0
sklearn==0.0.dist-info